requirements.txt
python-telegram-bot==21.10
gspread
oauth2client
python-dotenv
